Move OutputTruncationHelper to shared library and apply truncation in DLL module
Truncation now happens in PowerShellCommunication.NotifyResultReady() before
caching, so only the truncated preview crosses the named pipe — reducing pipe
transfer size and memory overhead. Full output is saved to a temp file on the
DLL side.

- Add PowerShell.MCP.Shared project with OutputTruncationHelper (threshold 15K)
- Remove OutputTruncationHelper from Proxy/Helpers (no longer needed there)
- Add NotifySilentResultReady() for small known-output internal paths
- Update project references and solution file
- Move tests to Tests/Unit/Shared/
