+++ PowerShell.MCP.sln
@@ -11,6 +11,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{0AB3BF05
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PowerShell.MCP.Tests", "Tests\PowerShell.MCP.Tests.csproj", "{115EE524-BDC8-496D-BF2B-E851A4E0CF19}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PowerShell.MCP.Shared", "PowerShell.MCP.Shared\PowerShell.MCP.Shared.csproj", "{E4A3F5C1-B0A2-4D6E-9F8B-7C2D1A0E3B5F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -57,6 +59,18 @@ Global
 		{115EE524-BDC8-496D-BF2B-E851A4E0CF19}.Release|x64.Build.0 = Release|Any CPU
 		{115EE524-BDC8-496D-BF2B-E851A4E0CF19}.Release|x86.ActiveCfg = Release|Any CPU
 		{115EE524-BDC8-496D-BF2B-E851A4E0CF19}.Release|x86.Build.0 = Release|Any CPU
+		{E4A3F5C1-B0A2-4D6E-9F8B-7C2D1A0E3B5F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E4A3F5C1-B0A2-4D6E-9F8B-7C2D1A0E3B5F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E4A3F5C1-B0A2-4D6E-9F8B-7C2D1A0E3B5F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{E4A3F5C1-B0A2-4D6E-9F8B-7C2D1A0E3B5F}.Debug|x64.Build.0 = Debug|Any CPU
+		{E4A3F5C1-B0A2-4D6E-9F8B-7C2D1A0E3B5F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E4A3F5C1-B0A2-4D6E-9F8B-7C2D1A0E3B5F}.Debug|x86.Build.0 = Debug|Any CPU
+		{E4A3F5C1-B0A2-4D6E-9F8B-7C2D1A0E3B5F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E4A3F5C1-B0A2-4D6E-9F8B-7C2D1A0E3B5F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E4A3F5C1-B0A2-4D6E-9F8B-7C2D1A0E3B5F}.Release|x64.ActiveCfg = Release|Any CPU
+		{E4A3F5C1-B0A2-4D6E-9F8B-7C2D1A0E3B5F}.Release|x64.Build.0 = Release|Any CPU
+		{E4A3F5C1-B0A2-4D6E-9F8B-7C2D1A0E3B5F}.Release|x86.ActiveCfg = Release|Any CPU
+		{E4A3F5C1-B0A2-4D6E-9F8B-7C2D1A0E3B5F}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
